# backend/openapi/uniflow-api.yaml
openapi: 3.0.3
info:
  title: Uniflow API
  version: 1.0.0
  description: >
    Specification OpenAPI des endpoints actuellement implementes dans le backend Uniflow.
servers:
  - url: http://localhost:8080
    description: Serveur local Spring Boot
...
paths:
...
  /api/events/orga/{organisationId}:
    get:
      tags: [Events]
      summary: Lister les evenements d'une organisation
      operationId: getEventsByOrganisation
      parameters:
        - in: path
          name: organisationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste des evenements de l'organisation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
...
  /api/inscriptions/{userId}:
    get:
      tags: [Inscriptions]
      summary: Lister les evenements auxquels un utilisateur est inscrit
      operationId: getUserInscriptions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Liste des inscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/inscriptions/event/{eventId}:
    get:
      tags: [Inscriptions]
      summary: Lister les participants d'un evenement
      operationId: getEventParticipants
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Liste des participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipantDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/inscriptions/{userId}/{eventId}:
    post:
      tags: [Inscriptions]
      summary: Inscrire un utilisateur a un evenement
      operationId: addInscription
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/EventId'
      responses:
        '201':
          description: Inscription ajoutee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Inscriptions]
      summary: Supprimer une inscription
      operationId: removeInscription
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Inscription supprimee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/organizers/create:
    post:
      tags: [Organizers]
      summary: Creer une organisation
      operationId: createOrganizer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganisationDTO'
      responses:
        '201':
          description: Organisation creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/organizers/{id}:
    get:
      tags: [Organizers]
      summary: Recuperer une organisation par identifiant
      operationId: getOrganizerById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Organisation trouvee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Organizers]
      summary: Modifier une organisation
      operationId: updateOrganizer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganisationDTO'
      responses:
        '200':
          description: Organisation modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Organizers]
      summary: Supprimer une organisation
      operationId: deleteOrganizer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Organisation supprimee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
    UserId:
      in: path
      name: userId
      required: true
      schema:
        type: integer
        format: int64
    EventId:
      in: path
      name: eventId
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Donnees invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentification requise ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Acces interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflit avec une ressource existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    EventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Conference Intelligence Artificielle
        description:
          type: string
          example: Decouvrez les dernieres avancees en intelligence artificielle.
        address:
          type: string
          example: Amphitheatre Themis - Campus Lyon 1
        beginDate:
          type: string
          format: date-time
          example: 2026-04-15T09:00:00
        endDate:
          type: string
          format: date-time
          example: 2026-04-15T17:00:00
        capacity:
          type: integer
          nullable: true
          example: 150
        status:
          type: string
          enum: [PUBLISHED, CLOSED, CANCELED]
          example: PUBLISHED
        price:
          type: number
          format: double
          example: 25.0
        organisationId:
          type: integer
          format: int64
          nullable: true
          example: 1
        organisationName:
          type: string
          nullable: true
          example: Club Informatique
        poleId:
          type: integer
          format: int64
          nullable: true
          example: 1
        poleName:
          type: string
          nullable: true
          example: Informatique
    OrganisationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: Club Informatique
        nomStructure:
          type: string
          example: Club Informatique
        email:
          type: string
          format: email
          nullable: true
          example: club.info@univ-lyon1.fr
        pole:
          type: string
          nullable: true
          example: Informatique
        poleId:
          type: integer
          format: int64
          nullable: true
          example: 1
        status:
          type: string
          example: Actif
    ParticipantDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nom:
          type: string
          example: Dupont
        prenom:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        eventId:
          type: integer
          format: int64
          example: 12
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation effectuee avec succes
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Identifiants invalides.